Possible fix to vm::poll() method.
After this commit, when the poll method failed to get the status of the VM, the WU was automatically canceled. This new
version allows the wrapper to try it 3 more times, waiting 5 seconds before running the next poll. If after 3 times the
status cannot be obtained, the WU will be aborted, otherwise the counter will be reset to zero.