Improved registry key read operation closes #9.
This commit should fix the bug #9. The operation now is taken, checking first if the Registry Key exists, then, getting the
size to read, and finally storing it in a variable. The process uses registry key handles, and cleans memory accordingly.

Additionally, the code has been simplified to set up the PATH env variable.