Added Port Forwarding, closes #12.
When the VM is created, the following port forwarding is set up:

Any connection to port 0.0.0.0:7859 will be forwarded to virtual_machine_IP:80.

Additionally, this commit adds a recovery system for starting the VM. A maximum amount of 4 attempts are allowed, after that
the work unit will be canceled the VM will be removed. The remove operation has been also simplified relying only in
VBoxManage unregistervm --delete command.