Multiple cores can be used in the VM. Closes #21.
The wrapper now detects how many cores has the computer (reported via BOINC
client) and according to the user preferences it creates a Virtual Machine with
the allowed number of cores. This will force the VM to have the same amount of
cores for the whole WU, while VirtualBox supports CPU hot plugging, however,
that scenario should be studied before deploying it.