Multi-core via command line. Closes #33
Previous multi-core code broke internal BOINC scheduling, as the VM will
be using more cores than BOINC thought. In order to run a multi-core
system in BOINC it is mandatory to do it via the server, so the BOINC
core client can know how many cores each WU is using.